Fix crash on skin reload when dialog was opened from a list control
Crash occurs because object contained by the smart pointer is destructed
before the function returns when the skin is reloaded. This fixes the
problem by avoiding additional copies of the object persisting past the
skin reload. Skin reload happens just prior to the function returning, so
destructing the object inside the call doesn't cause a crash.