Peripheral API v1.4.0 - Offset discrete D-pads
Discrete D-pads are D-pads that use analog axes with discrete values
instead of digital buttons.

Normally, discrete D-pads are centered at 0.0, travel to -1.0 when left is
pressed, and 1.0 when right is pressed.

On Windows, discrete D-pads have been observed that start centered at 0.498,
travel to 0.0 when left is pressed, and 1.0 when right is pressed.