gui/texture/gl(es): Apply state block before rendering
There are OpenGL errors when glVertexAttribute is called in
CGUITextureGL::End after starting up media (only tested on RetroPlayer,
but it probably happens on VideoPlayer as well).

The cause of the error is that there's no VAO bound (or rather, VAO 0 is
bound).
Not sure why rendering has worked up to this point.
This change asks the rendering context to apply any state necessary
before rendering (part of which is binding the VAO).

It also fixes a GUI freeze issue on RetroPlayer.

This is tested with GL core 4.6 profile, not sure about older or
compatibility profiles.

Running apitrace doesn't show the related GL errors after this change.